Move FtLibraryHandler to a separate file.
