+++ rts/Rendering/CMakeLists.txt
@@ -138,6 +138,7 @@ set(sources_engine_Rendering
 		"${CMAKE_CURRENT_SOURCE_DIR}/VerticalSync.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/WorldDrawer.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Fonts/FontHandler.cpp"
+		"${CMAKE_CURRENT_SOURCE_DIR}/Fonts/FtLibraryHandler.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Fonts/CFontTexture.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Fonts/glFont.cpp"
 		"${CMAKE_CURRENT_SOURCE_DIR}/Fonts/glFontRenderer.cpp"
