+++ rts/Rendering/Fonts/CFontTexture.cpp
@@ -5,6 +5,7 @@
 #include "glFontRenderer.h"
 #include "FontHandler.h"
 #include "FontLogSection.h"
+#include "FtLibraryHandler.h"
 
 #include <cstring> // for memset, memcpy
 #include <string>
@@ -93,253 +94,6 @@ static spring::unordered_map<SizedFontKey, TimestampedFont> pinnedRecentFonts;
 #include "NonPrintableSymbols.inl"
 
 
-#ifndef HEADLESS
-class FtLibraryHandler {
-public:
-	FtLibraryHandler()
-		: config(nullptr)
-		, lib(nullptr)
-		#ifdef USE_FONTCONFIG
-		, gameFontSet(nullptr)
-		, basePattern(nullptr)
-		#endif // USE_FONTCONFIG
-	{
-		const FT_Error error = FT_Init_FreeType(&lib);
-
-		if (error != 0) {
-			FT_Int version[3];
-			FT_Library_Version(lib, &version[0], &version[1], &version[2]);
-
-			std::string err = fmt::sprintf("[%s] FT_Init_FreeType failure (version %d.%d.%d) \"%s\"",
-						       __func__, version[0], version[1], version[2], GetFTError(error));
-			throw std::runtime_error(err);
-		}
-	}
-
-	~FtLibraryHandler() {
-		pinnedRecentFonts.clear();
-		FT_Done_FreeType(lib);
-
-		#ifdef USE_FONTCONFIG
-		if (!config)
-			return;
-
-		FcConfigDestroy(config);
-		if (gameFontSet) {
-			FcFontSetDestroy(gameFontSet);
-		}
-		if (basePattern) {
-			FcPatternDestroy(basePattern);
-		}
-		FcFini();
-		config = nullptr;
-		#endif
-	}
-
-	bool InitFontconfig(bool console) {
-		#ifdef USE_FONTCONFIG
-		auto LOG_MSG = [console](const std::string& fmt, bool isError, auto&&... args) {
-			if (console) {
-				std::string fmtNL = fmt + "\n";
-				printf(fmtNL.c_str(), args...);
-			}
-			else {
-				if (isError) {
-					LOG_L(L_ERROR, fmt.c_str(), args...);
-				}
-				else {
-					LOG(fmt.c_str(), args...);
-				}
-			}
-		};
-
-		if (!UseFontConfig())
-			return false;
-
-		{
-			std::string msg = fmt::sprintf("%s::FontConfigInit (version %d.%d.%d)", __func__, FC_MAJOR, FC_MINOR, FC_REVISION);
-			ScopedOnceTimer timer(msg);
-			ZoneScopedNC("FtLibraryHandler::FontConfigInit", tracy::Color::Purple);
-
-			FcBool res;
-			std::string errprefix = fmt::sprintf("[%s] Fontconfig(version %d.%d.%d) failed to initialize", __func__, FC_MAJOR, FC_MINOR, FC_REVISION);
-
-			// init configuration
-			FcConfigEnableHome(FcFalse);
-			config = FcConfigCreate();
-
-			// we cant directly use the usual fontconfig methods because those won't let us have both first our cache
-			// and system fonts included. also linux actually has system config files that can be used by fontconfig.
-
-			#ifdef _WIN32
-			static constexpr auto winFontPath = "%WINDIR%\\fonts";
-			const int neededSize = ExpandEnvironmentStrings(winFontPath, nullptr, 0);
-			std::vector <char> osFontsDir (neededSize);
-			ExpandEnvironmentStrings(winFontPath, osFontsDir.data(), osFontsDir.size());
-
-			static constexpr const char* configFmt = R"(<fontconfig><dir>%s</dir><cachedir>fontcache</cachedir></fontconfig>)";
-			const std::string configFmtVar = fmt::sprintf(configFmt, osFontsDir.data());
-			#else
-			const std::string configFmtVar = R"(<fontconfig><cachedir>fontcache</cachedir></fontconfig>)";
-			#endif
-
-			#ifdef _WIN32
-			// Explicitly set the config with xml for windows.
-			res = FcConfigParseAndLoadFromMemory(config, reinterpret_cast<const FcChar8*>(configFmtVar.c_str()), FcTrue);
-			#else
-			// Load system configuration (passing 0 here so fc will use the default os config file if possible).
-			res = FcConfigParseAndLoad(config, 0, true);
-			#endif
-			if (res) {
-				#ifndef _WIN32
-				// add local cache after system config for linux
-				FcConfigParseAndLoadFromMemory(config, reinterpret_cast<const FcChar8*>(configFmtVar.c_str()), FcTrue);
-				#endif
-
-				LOG_MSG("[%s] Using Fontconfig light init", false, __func__);
-
-				// build system fonts
-				res = FcConfigBuildFonts(config);
-				if (!res) {
-					LOG_MSG("%s fonts", true, errprefix.c_str());
-					InitFailed();
-					return false;
-				}
-			} else {
-				// Can't load step by step to use our cache, so retry with general
-				// fontconfig init method, that has a few extra fallbacks.
-
-				// Init everything. Normally this would be enough, but the method before
-				// accounts for situations where system config is borked due to incompatible
-				// lib and system config files.
-				FcConfig *fcConfig = FcInitLoadConfigAndFonts();
-				if (fcConfig) {
-					FcConfigDestroy(config); // release previous config
-					config = fcConfig;
-
-					// add our cache at the back of the new config.
-					FcConfigParseAndLoadFromMemory(config, reinterpret_cast<const FcChar8*>(configFmtVar.c_str()), FcTrue);
-				} else {
-					LOG_MSG("%s config and fonts. No system fallbacks will be available", false, errprefix.c_str());
-				}
-			}
-
-			gameFontSet = FcFontSetCreate();
-			basePattern = FcPatternCreate();
-
-			// init app fonts dir
-			res = FcConfigAppFontAddDir(config, reinterpret_cast<const FcChar8*>("fonts"));
-			if (!res) {
-				LOG_MSG("%s font dir", true, errprefix.c_str());
-				InitFailed();
-				return false;
-			}
-
-			// print cache dirs
-			auto dirs = FcConfigGetCacheDirs(config);
-			FcStrListFirst(dirs);
-			for (FcChar8* dir = FcStrListNext(dirs); dir != nullptr; dir = FcStrListNext(dirs)) {
-				LOG_MSG("[%s] Using Fontconfig cache dir \"%s\"", false, __func__, dir);
-			}
-			FcStrListDone(dirs);
-		}
-
-		#endif // USE_FONTCONFIG
-
-		return true;
-	}
-
-	void InitFailed() {
-		FcConfigDestroy(config);
-		FcFini();
-		config = nullptr;
-	}
-	static bool InitSingletonFontconfig(bool console) { return singleton->InitFontconfig(console); }
-
-	static bool UseFontConfig() { return fontHandler.useFontConfigLib; }
-
-	#ifdef USE_FONTCONFIG
-	// command-line CheckGenFontConfigFull invocation checks
-	static bool CheckFontConfig() { return (UseFontConfig() && FcConfigUptoDate(GetFCConfig())); }
-	#else
-
-	static bool CheckFontConfig() { return false; }
-	static bool CheckGenFontConfig(bool fromCons) { return false; }
-	#endif
-
-	static FT_Library& GetLibrary() {
-		if (singleton == nullptr)
-			singleton = std::make_unique<FtLibraryHandler>();
-
-		return singleton->lib;
-	};
-	static FcConfig* GetFCConfig() {
-		if (singleton == nullptr)
-			singleton = std::make_unique<FtLibraryHandler>();
-
-		return singleton->config;
-	}
-	static inline bool CanUseFontConfig() {
-		return GetFCConfig() != nullptr;
-	}
-	#ifdef USE_FONTCONFIG
-	static FcFontSet *GetGameFontSet() {
-		return singleton->gameFontSet;
-	}
-	static FcPattern *GetBasePattern() {
-		return singleton->basePattern;
-	}
-	static void ClearGameFontSet() {
-		FcFontSetDestroy(singleton->gameFontSet);
-		singleton->gameFontSet = FcFontSetCreate();
-	}
-	static void ClearBasePattern() {
-		FcPatternDestroy(singleton->basePattern);
-		singleton->basePattern = FcPatternCreate();
-	}
-	static bool GetSearchSystemFonts() {
-		return fontHandler.searchSystemFonts;
-	}
-	static bool GetSearchFontAttributes() {
-		return fontHandler.searchFontAttributes;
-	}
-	static bool GetSearchApplySubstitutions() {
-		return fontHandler.searchApplySubstitutions;
-	}
-	#endif
-private:
-	FcConfig* config;
-	FT_Library lib;
-	#ifdef USE_FONTCONFIG
-	FcFontSet *gameFontSet;
-	FcPattern *basePattern;
-	#endif
-
-	static inline std::unique_ptr<FtLibraryHandler> singleton = nullptr;
-};
-#endif
-
-
-
-void FtLibraryHandlerProxy::InitFtLibrary()
-{
-	RECOIL_DETAILED_TRACY_ZONE;
-#ifndef HEADLESS
-	FtLibraryHandler::GetLibrary();
-#endif
-}
-
-bool FtLibraryHandlerProxy::InitFontconfig(bool console)
-{
-	RECOIL_DETAILED_TRACY_ZONE;
-#ifndef HEADLESS
-	return FtLibraryHandler::InitSingletonFontconfig(console);
-#else
-	return false;
-#endif
-}
-
-
 /*******************************************************************************/
 /*******************************************************************************/
 /*******************************************************************************/
@@ -394,7 +148,7 @@ static std::shared_ptr<FontFace> LoadFontFace(const std::string& fontfile)
 
 	// load the font
 	FT_Face face_ = nullptr;
-	FT_Error error = FT_New_Memory_Face(FtLibraryHandler::GetLibrary(), fontMem.get()->data(), filesize, 0, &face_);
+	FT_Error error = FtLibraryHandlerProxy::NewMemoryFace(fontMem.get()->data(), filesize, 0, &face_);
 	auto face = spring::ScopedResource(
 		face_,
 		[](FT_Face f) { if (f) FT_Done_Face(f); }
@@ -462,7 +216,7 @@ static std::shared_ptr<FontFace> GetFontForCharacters(const std::vector<char32_t
 	if (characters.empty())
 		return nullptr;
 
-	if (!FtLibraryHandler::CanUseFontConfig())
+	if (!FtLibraryHandlerProxy::CanUseFontConfig())
 		return nullptr;
 
 	// create list of wanted characters
@@ -477,7 +231,7 @@ static std::shared_ptr<FontFace> GetFontForCharacters(const std::vector<char32_t
 
 	// create properties of the wanted font starting from our priorities pattern.
 	auto pattern = spring::ScopedResource(
-		FcPatternDuplicate(FtLibraryHandler::GetBasePattern()),
+		FcPatternDuplicate(FtLibraryHandlerProxy::GetBasePattern()),
 		[](FcPattern* p) { if (p) FcPatternDestroy(p); }
 	);
 
@@ -507,7 +261,7 @@ static std::shared_ptr<FontFace> GetFontForCharacters(const std::vector<char32_t
 			[](FcBlanks* b) { if (b) FcBlanksDestroy(b); }
 		);
 
-		if (FtLibraryHandler::GetSearchFontAttributes()) {
+		if (fontHandler.searchFontAttributes) {
 			auto origPattern = spring::ScopedResource(
 				FcFreeTypeQueryFace(origFace, ftname, 0, blanks),
 				[](FcPattern* p) { if (p) FcPatternDestroy(p); }
@@ -536,7 +290,7 @@ static std::shared_ptr<FontFace> GetFontForCharacters(const std::vector<char32_t
 	}
 
 	FcDefaultSubstitute(pattern);
-	if (FtLibraryHandler::GetSearchApplySubstitutions() && !FcConfigSubstitute(FtLibraryHandler::GetFCConfig(), pattern, FcMatchPattern))
+	if (fontHandler.searchApplySubstitutions && !FcConfigSubstitute(FtLibraryHandlerProxy::GetFCConfig(), pattern, FcMatchPattern))
 	{
 		return nullptr;
 	}
@@ -545,12 +299,12 @@ static std::shared_ptr<FontFace> GetFontForCharacters(const std::vector<char32_t
 	typedef std::unique_ptr<FcFontSet, decltype(&FcFontSetDestroy)> ScopedFcFontSet;
 
 	int nFonts = 0;
-	bool loadMore = FtLibraryHandler::GetSearchSystemFonts();
+	bool loadMore = fontHandler.searchSystemFonts;
 	FcResult res;
 
 	// first search game fonts
-	FcFontSet *sets[] = { FtLibraryHandler::GetGameFontSet() };
-	ScopedFcFontSet fs(FcFontSetSort(FtLibraryHandler::GetFCConfig(), sets, 1, pattern, FcFalse, nullptr, &res), &FcFontSetDestroy);
+	FcFontSet *sets[] = { FtLibraryHandlerProxy::GetGameFontSet() };
+	ScopedFcFontSet fs(FcFontSetSort(FtLibraryHandlerProxy::GetFCConfig(), sets, 1, pattern, FcFalse, nullptr, &res), &FcFontSetDestroy);
 
 	if (fs != nullptr && res == FcResultMatch)
 		nFonts = fs->nfont;
@@ -560,7 +314,7 @@ static std::shared_ptr<FontFace> GetFontForCharacters(const std::vector<char32_t
 	while (i < nFonts || loadMore) {
 		if (i == nFonts) {
 			// now search system fonts
-			fs = ScopedFcFontSet(FcFontSort(FtLibraryHandler::GetFCConfig(), pattern, FcFalse, nullptr, &res), &FcFontSetDestroy);
+			fs = ScopedFcFontSet(FcFontSort(FtLibraryHandlerProxy::GetFCConfig(), pattern, FcFalse, nullptr, &res), &FcFontSetDestroy);
 			if (fs == nullptr || res != FcResultMatch)
 				return nullptr;
 			loadMore = false;
@@ -762,10 +516,10 @@ CFontTexture::~CFontTexture()
 bool CFontTexture::AddFallbackFont(const std::string& fontfile)
 {
 #if defined(USE_FONTCONFIG) && !defined(HEADLESS)
-	if (!FtLibraryHandler::CanUseFontConfig())
+	if (!FtLibraryHandlerProxy::CanUseFontConfig())
 		return false;
 
-	FcFontSet *set = FtLibraryHandler::GetGameFontSet();
+	FcFontSet *set = FtLibraryHandlerProxy::GetGameFontSet();
 
 	// Check if font already loaded
 	for (int i=0; set && i < set->nfont; ++i) {
@@ -798,12 +552,12 @@ bool CFontTexture::AddFallbackFont(const std::string& fontfile)
 		return false;
 	}
 	// needed?:
-	//FcConfigSubstitute(FtLibraryHandler::GetFCConfig(), pattern, FcMatchScan);
+	//FcConfigSubstitute(FtLibraryHandlerProxy::GetFCConfig(), pattern, FcMatchScan);
 
 	// Add to priority fonts pattern
 	FcChar8* family = nullptr;
 	if (FcPatternGetString( pattern, FC_FAMILY , 0, &family ) == FcResultMatch) {
-		FcPattern *basePattern = FtLibraryHandler::GetBasePattern();
+		FcPattern *basePattern = FtLibraryHandlerProxy::GetBasePattern();
 		FcPatternAddString(basePattern, FC_FAMILY, family);
 	} else {
 		LOG_L(L_WARNING, "[%s] could not add priority for %s", __func__, fontfile.c_str());
@@ -826,11 +580,11 @@ void CFontTexture::ClearFallbackFonts()
 {
 	pinnedRecentFonts.clear();
 #if defined(USE_FONTCONFIG) && !defined(HEADLESS)
-	if (!FtLibraryHandler::CanUseFontConfig())
+	if (!FtLibraryHandlerProxy::CanUseFontConfig())
 		return;
 
-	FtLibraryHandler::ClearBasePattern();
-	FtLibraryHandler::ClearGameFontSet();
+	FtLibraryHandlerProxy::ClearBasePattern();
+	FtLibraryHandlerProxy::ClearGameFontSet();
 
 	needsClearGlyphs = true;
 #endif
@@ -931,6 +685,9 @@ void CFontTexture::InitFonts()
 void CFontTexture::KillFonts()
 {
 	RECOIL_DETAILED_TRACY_ZONE;
+
+	pinnedRecentFonts.clear();
+
 	// check unused fonts
 	std::erase_if(allFonts, [](std::weak_ptr<CFontTexture> item) { return item.expired(); });
